+++ internal/model/transaction.go
@@ -7,7 +7,7 @@ import (
 type Transaction struct {
 	ID         int       `json:"id"`
 	CategoryID int       `json:"category_id"`
-	Amount     float64       `json:"amount"`
+	Amount     int64       `json:"amount"`
 	CreatedAt  time.Time `json:"created_at"`
 	UpdatedAt  time.Time `json:"updated_at"`
 }
