+++ internal/usecase/transaction/add/add.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 )
 
-func (u *Usecase) Add(ctx context.Context, categoryID int, amount float64) error {
+func (u *Usecase) Add(ctx context.Context, categoryID int, amount int64) error {
 	if err := u.transactionRepo.AddTransaction(ctx, categoryID, amount); err != nil {
 		return fmt.Errorf("Failed to add transaction: %w", err)
 	}
